backend/requirements.txt
fastapi>=0.100.0
uvicorn[standard]>=0.23.0
pydantic>=1.10.0
pydantic-settings>=2.0.0
fpdf2
qrcode
schedule
requests
openai
python-dotenv>=0.20.0
sqlalchemy>=1.4.0
pillow
# Autenticación y Seguridad
pyjwt>=2.8.0
passlib[bcrypt]>=1.7.4
python-jose[cryptography]>=3.3.0
python-multipart>=0.0.6
# Rate Limiting y Hardening
slowapi>=0.1.9
bleach>=6.0.0

# Testing
pytest>=9.0.0
pytest-cov>=7.0.0
pytest-asyncio>=1.3.0
responses>=0.25.0
faker>=40.0.0
httpx>=0.28.0
PyPDF2>=3.0.0


